Remove filesystem dependency on project
