+++ crates/cairo-lang-filesystem/Cargo.toml
@@ -10,7 +10,6 @@ description = "Virtual filesystem for the compiler."
 cairo-lang-utils = { path = "../cairo-lang-utils", version = "0.1.0" }
 cairo-lang-debug = { path = "../cairo-lang-debug", version = "0.1.0" }
 path-clean.workspace = true
-cairo-lang-project = { path = "../cairo-lang-project", version = "0.1.0" }
 salsa.workspace = true
 smol_str.workspace = true
