+++ crates/cairo-lang-project/Cargo.toml
@@ -11,6 +11,7 @@ serde.workspace = true
 smol_str.workspace = true
 thiserror.workspace = true
 toml.workspace = true
+cairo-lang-filesystem = { path = "../cairo-lang-filesystem", version = "0.1.0" }
 
 [dev-dependencies]
 indoc.workspace = true
